+++ src/JK.Common/CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Unless otherwise noted, all changes made by [@jeremyknight-me](https://github.com/jeremyknight-me).
 
+## 3.0.1
+
+### Added
+
+- Last alias for Right on StringHelper
+
+### Changed
+
+- Changed StringHelper to static class
+
 ## 3.0.0
 
 ### Added